Revert "Converts to use gitops provider (#45)" (#46)
This reverts commit 038377bc7ccd08eab3be4553b8b1d10f1c29d325.

+++ .github/scripts/validate-deploy.sh
@@ -1,23 +1,9 @@
 #!/usr/bin/env bash
 
-SCRIPT_DIR=$(cd $(dirname "$0"); pwd -P)
-
 GIT_REPO=$(cat git_repo)
 GIT_TOKEN=$(cat git_token)
 
-BIN_DIR=$(cat .bin_dir)
-
-export PATH="${BIN_DIR}:${PATH}"
-
 export KUBECONFIG=$(cat .kubeconfig)
-
-source "${SCRIPT_DIR}/validation-functions.sh"
-
-NAMESPACE=$(jq -r '.name // "gitops-namespace"' gitops-output.json)
-BRANCH=$(jq -r '.branch // "main"' gitops-output.json)
-SERVER_NAME=$(jq -r '.server_name // "default"' gitops-output.json)
-LAYER=$(jq -r '.layer_dir // "2-services"' gitops-output.json)
-TYPE=$(jq -r '.type // "base"' gitops-output.json)
 
 mkdir -p .testrepo
 
@@ -27,10 +13,61 @@ cd .testrepo || exit 1
 
 find . -name "*"
 
-validate_gitops_ns_content "${NAMESPACE}"
+SERVER_NAME="default"
+NAMESPACE="gitops-namespace"
+
+if [[ ! -f "argocd/1-infrastructure/cluster/${SERVER_NAME}/base/namespace-${NAMESPACE}.yaml" ]]; then
+  echo "Argocd config missing: argocd/1-infrastructure/cluster/${SERVER_NAME}/base/namespace-${NAMESPACE}.yaml"
+  exit 1
+fi
+
+echo "Printing argocd/1-infrastructure/cluster/${SERVER_NAME}/base/namespace-${NAMESPACE}.yaml"
+cat "argocd/1-infrastructure/cluster/${SERVER_NAME}/base/namespace-${NAMESPACE}.yaml"
+
+if [[ ! -f "argocd/1-infrastructure/cluster/${SERVER_NAME}/kustomization.yaml" ]]; then
+  echo "Argocd config missing: argocd/1-infrastructure/cluster/${SERVER_NAME}/kustomization.yaml"
+  exit 1
+fi
+
+echo "Printing argocd/1-infrastructure/cluster/${SERVER_NAME}/kustomization.yaml"
+cat "argocd/1-infrastructure/cluster/${SERVER_NAME}/kustomization.yaml"
+
+if [[ ! -f "payload/1-infrastructure/namespace/${NAMESPACE}/namespace/ns.yaml" ]]; then
+  echo "Payload missing: payload/1-infrastructure/namespace/${NAMESPACE}/namespace/ns.yaml"
+  exit 1
+fi
+
+echo "Printing payload/1-infrastructure/namespace/${NAMESPACE}/namespace/ns.yaml"
+cat "payload/1-infrastructure/namespace/${NAMESPACE}/namespace/ns.yaml"
+
+if [[ ! -f "payload/1-infrastructure/namespace/${NAMESPACE}/namespace/rbac.yaml" ]]; then
+  echo "Payload missing: payload/1-infrastructure/namespace/${NAMESPACE}/namespace/rbac.yaml"
+  exit 1
+fi
+
+echo "Printing payload/1-infrastructure/namespace/${NAMESPACE}/namespace/rbac.yaml"
+cat "payload/1-infrastructure/namespace/${NAMESPACE}/namespace/rbac.yaml"
 
 cd ..
 rm -rf .testrepo
 
-check_k8s_namespace "${NAMESPACE}"
-check_k8s_resource "${NAMESPACE}" rolebinding argocd-admin
+count=0
+until kubectl get namespace "${NAMESPACE}" 1> /dev/null 2> /dev/null || [[ $count -eq 20 ]]; do
+  echo "Waiting for namespace: ${NAMESPACE}"
+  count=$((count + 1))
+  sleep 15
+done
+
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for namespace: ${NAMESPACE}"
+  exit 1
+else
+  echo "Found namespace: ${NAMESPACE}. Sleeping for 30 seconds to wait for everything to settle down"
+  sleep 30
+fi
+
+if ! kubectl get rolebinding -n "${NAMESPACE}" argocd-admin 1> /dev/null 2> /dev/null; then
+  echo "Unable to find rolebinding: ${NAMESPACE}/argocd-admin"
+  kubectl get rolebinding -n "${NAMESPACE}"
+  exit 1
+fi
